sharedrandom の責務分割
