+++ src/hooks/useGameBoardLogic.ts
@@ -12,7 +12,8 @@ import { uuid } from '../utils/helpers';
 import * as CardLogic from '../utils/cardLogic';
 import { canImportDeck } from '../utils/gameRules';
 import { applyGameSyncEvent } from '../utils/gameSyncReducer';
-import { flipSharedCoin, formatSharedUiMessage, getSharedActorLabel, rollSharedDie } from '../utils/sharedRandom';
+import { flipSharedCoin, rollSharedDie } from '../utils/sharedRandom';
+import { formatSharedUiMessage, getSharedActorLabel } from '../utils/sharedUiMessage';
 import { createEventDeduper } from '../utils/eventDeduper';
 import { type CardStatLookup } from '../utils/cardStats';
 import { type CardDetailLookup } from '../utils/cardDetails';
